requirements.txt
# Tenant Lease Analyzer - Requirements

# Core AI/LLM
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-community==0.0.20
anthropic==0.18.1

# PDF Processing
pypdf==4.0.1
PyPDF2==3.0.1

# Web Interface
streamlit==1.31.0

# Data Processing
pandas==2.2.0

# Environment Management
python-dotenv==1.0.1

# Optional: For production deployment
gunicorn==21.2.0
